+++ data/footer.php
@@ -1,0 +1,12 @@
+<!-- Fixed Footer -->
+<footer class="fixed-bottom bg-primary text-white">
+    <div class="container text-center small">
+        <div>Copyright © 2025 Lee Bussy [AA0NT].</div>
+
+        <div>
+            WSPR Version: <span id="wspr-version">Loading version...</span>
+        </div>
+
+        <div>Licensed under the MIT License</div>
+    </div>
+</footer>
