+++ data/site.js.includes.php
@@ -1,0 +1,13 @@
+<!-- jQuery 3.7.1 -->
+<script
+    src="https://code.jquery.com/jquery-3.7.1.min.js"
+    integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo="
+    crossorigin="anonymous">
+</script>
+
+<!-- Bootstrap Bundle JS (Includes Popper) -->
+<script
+    src="https://bootswatch.com/_vendor/bootstrap/dist/js/bootstrap.bundle.min.js"
+    integrity="sha384-k6d4wzSIapyDyv1kpU366/PK5hCdSbCRGRCMv+eplOQJWyd1fbcAu9OCUj5zNLiq"
+    crossorigin="anonymous">
+</script>
