+++ data/template.php
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en" data-bs-theme="light">
+
+<head>
+    <!-- Bootswatch, Boostrap, and Fontawesome, included here: -->
+    <?php require_once 'header.php'; ?>
+
+    <!-- This page's css -->
+    <link rel="stylesheet" href="site.css" />
+
+    <!-- This page's css -->
+    <link rel="stylesheet" href="template.css" />
+</head>
+
+<body>
+    <?php require_once 'navbar.php'; ?>
+
+    <!-- Main Content Goes Here -->
+
+    <!-- Static page footer -->
+    <?php require_once 'footer.php'; ?>
+
+    <!-- jQuery and Bootswatch -->
+    <?php require_once 'site.js.includes.php'; ?>
+
+    <!-- Main JavaScript -->
+    <script src="site.js"></script>
+
+    <!-- This page's JavaScript -->
+    <script src="template.js"></script>
+</body>
+
+</html>
